feat:  migrations typescript
기존 자바스크립트로 되어있던 프로젝트를 타입스크립트로 마이그레이션 하였습니다
