+++ src/components/PageList.tsx
@@ -1,12 +1,16 @@
-import React from "react";
-import styled from "styled-components";
+import React from 'react';
+import styled from 'styled-components';
+
+interface StyleProps {
+  active?: boolean;
+}
 
 const PageListStyle = styled.div`
   margin-bottom: 20px;
   text-align: center;
 `;
 
-const Page = styled.button`
+const Page = styled.button<StyleProps>`
   padding: 0.375rem 0.75rem;
   border-radius: 0.25rem;
   font-size: 1rem;
@@ -26,7 +30,7 @@ function PageList() {
 
   pageArray.push(
     // 임시로 페이지 하나만 설정했습니다.
-    <Page key="1">1</Page>
+    <Page key="1">1</Page>,
   );
 
   return <PageListStyle>{pageArray}</PageListStyle>;
