+++ src/containers/FormContainer.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import Form from "../components/Form";
+import React from 'react';
+import Form from '../components/Form';
 
 function FormContainer() {
   return <Form />;
